move sign-claim to stf/cli and fix wrong descriptions for CLI functions
